+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 - Fix: WASD controls now always in-sync with camera position
 - Fix: broken detach/attach functionality for some activities now resolved.
 - Fix: Webview disposal revised and corrected.
+- Fix: Prevent duplication of route IDs when loading source code.
 - New navigation controls to move camera in scene with zoom in/out function.
 - Scanning of header/property variables improved (includes parent/child routes)
+- New HTTP/S endpoint using new configuration system
+- All endpoints (to/toD) upgraded to use new configuration system
+- Starting activities 'direct' and 'timer' now configurable.
 
 
 ## 0.10.0